Added temp audio classification experimentation
No meaningful embeddings are able to be generated with current network, so experimenting with classification to see if this task is possible with current data. This is primarily to gauge data problems, as this task is simpler than embedding generation (while being somewhat similar), so it will be easier to debug the issues.